+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetDriver.cpp
@@ -272,6 +272,24 @@ void USpatialNetDriver::CreateAndInitializeCoreClasses()
 	if (IsServer())
 	{
 		EntityPool->Init(this, &TimerManager);
+
+		// Create a entity representing the server worker.
+		// This ensures that even if a server-worker isn't authoritative over anything, it will be able to
+		// receive updates for the GSM.
+		const FString WorkerId = FString(TEXT("workerId:")) + Connection->GetWorkerId();
+		const WorkerRequirementSet WorkerIdPermission{ { WorkerId } };
+
+		WriteAclMap ComponentWriteAcl;
+		ComponentWriteAcl.Add(SpatialConstants::POSITION_COMPONENT_ID, WorkerIdPermission);
+		ComponentWriteAcl.Add(SpatialConstants::METADATA_COMPONENT_ID, WorkerIdPermission);
+		ComponentWriteAcl.Add(SpatialConstants::ENTITY_ACL_COMPONENT_ID, WorkerIdPermission);
+
+		TArray<Worker_ComponentData> Components;
+		Components.Add(improbable::Position().CreatePositionData());
+		Components.Add(improbable::Metadata(WorkerId).CreateMetadataData());
+		Components.Add(improbable::EntityAcl(WorkerIdPermission, ComponentWriteAcl).CreateEntityAclData());
+
+		Connection->SendCreateEntityRequest(MoveTemp(Components), nullptr);
 	}
 }