+++ CHANGELOG.md
@@ -58,6 +58,7 @@ In addition to all of the updates from Improbable, this release includes x impro
 ### Features:
 - If using Spatial Networking, the blueprint world selector dropdown now shows the worker ID for each world for easier debugging.
 - Added toggle for enabling/disabling offloading.
+- Entities are now generated for each server-worker type to ensure that workers that initially are not authoritative over any components will receive updates for the GSM entity.
 
 ## [`0.4.1`](https://github.com/spatialos/UnrealGDK/releases/tag/0.4.1) - 2019-05-01